+++ src/ffi.rs
@@ -6,7 +6,7 @@ pub struct kernel_stack_alloc_data {
 }
 #[no_mangle]
 pub static mut kernel_stack_alloc: kernel_stack_alloc_data = kernel_stack_alloc_data {
-    data: [[0 as u8; BIT!(CONFIG_KERNEL_STACK_BITS)]; CONFIG_MAX_NUM_NODES],
+    data: [[0_u8; BIT!(CONFIG_KERNEL_STACK_BITS)]; CONFIG_MAX_NUM_NODES],
 };
 #[cfg(feature = "ENABLE_SMP")]
 /// This function is used to map the core.
